+++ solr/core/src/java/org/apache/solr/update/DirectUpdateHandler2.java
@@ -50,6 +50,9 @@ import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.core.SolrConfig.UpdateHandlerInfo;
 import org.apache.solr.core.SolrCore;
+import org.apache.solr.core.SolrInfoBean;
+import org.apache.solr.metrics.SolrDelegateRegistryMetricsContext;
+import org.apache.solr.metrics.SolrMetricManager;
 import org.apache.solr.metrics.SolrMetricProducer;
 import org.apache.solr.metrics.SolrMetricsContext;
 import org.apache.solr.request.LocalSolrQueryRequest;
@@ -62,6 +65,7 @@ import org.apache.solr.search.FunctionRangeQuery;
 import org.apache.solr.search.QParser;
 import org.apache.solr.search.QueryUtils;
 import org.apache.solr.search.SolrIndexSearcher;
+import org.apache.solr.search.SolrSearcherRequirementDetector;
 import org.apache.solr.search.SyntaxError;
 import org.apache.solr.search.function.ValueSourceRangeFilter;
 import org.apache.solr.util.RefCounted;
@@ -75,6 +79,8 @@ import org.slf4j.LoggerFactory;
  */
 public class DirectUpdateHandler2 extends UpdateHandler
     implements SolrCoreState.IndexWriterCloser, SolrMetricProducer {
+
+  private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   private static final int NO_FILE_SIZE_UPPER_BOUND_PLACEHOLDER = -1;
 
@@ -102,6 +108,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
   protected final CommitTracker softCommitTracker;
 
   protected boolean commitWithinSoftCommit;
+
   /**
    * package access for testing
    *
@@ -111,10 +118,8 @@ public class DirectUpdateHandler2 extends UpdateHandler
     this.commitWithinSoftCommit = value;
   }
 
-  private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-
   public DirectUpdateHandler2(SolrCore core) {
-    super(core);
+    super(core, null, false);
 
     solrCoreState = core.getSolrCoreState();
 
@@ -152,10 +157,13 @@ public class DirectUpdateHandler2 extends UpdateHandler
       commitWithinSoftCommit = false;
       commitTracker.setOpenSearcher(true);
     }
+    if (ulog != null) {
+      initUlog(true);
+    }
   }
 
   public DirectUpdateHandler2(SolrCore core, UpdateHandler updateHandler) {
-    super(core, updateHandler.getUpdateLog());
+    super(core, updateHandler.getUpdateLog(), false);
     solrCoreState = core.getSolrCoreState();
 
     UpdateHandlerInfo updateHandlerInfo = core.getSolrConfig().getUpdateHandlerInfo();
@@ -186,17 +194,29 @@ public class DirectUpdateHandler2 extends UpdateHandler
 
     commitWithinSoftCommit = updateHandlerInfo.commitWithinSoftCommit;
 
-    UpdateLog existingLog = updateHandler.getUpdateLog();
-    if (this.ulog != null && this.ulog == existingLog) {
+    if (ulog != null) {
       // If we are reusing the existing update log, inform the log that its update handler has
       // changed. We do this as late as possible.
-      this.ulog.init(this, core);
+      // TODO: not sure _why_ we "do this as late as possible". Consider simplifying by
+      //  moving `ulog.init(UpdateHandler, SolrCore)` entirely into the `UpdateHandler` ctor,
+      //  avoiding the need for the `UpdateHandler(SolrCore, UpdateLog, boolean)` ctor
+      //  (with the extra boolean `initUlog` param).
+      initUlog(ulog != updateHandler.getUpdateLog());
     }
   }
 
   @Override
   public void initializeMetrics(SolrMetricsContext parentContext, String scope) {
-    solrMetricsContext = parentContext.getChildContext(this);
+    if (core.getSolrConfig().getUpdateHandlerInfo().aggregateNodeLevelMetricsEnabled) {
+      this.solrMetricsContext =
+          new SolrDelegateRegistryMetricsContext(
+              parentContext.getMetricManager(),
+              parentContext.getRegistryName(),
+              SolrMetricProducer.getUniqueMetricTag(this, parentContext.getTag()),
+              SolrMetricManager.getRegistryName(SolrInfoBean.Group.node));
+    } else {
+      this.solrMetricsContext = parentContext.getChildContext(this);
+    }
     commitCommands = solrMetricsContext.meter("commits", getCategory().toString(), scope);
     solrMetricsContext.gauge(
         () -> commitTracker.getCommitCount(), true, "autoCommits", getCategory().toString(), scope);
@@ -366,13 +386,12 @@ public class DirectUpdateHandler2 extends UpdateHandler
       }
 
       if ((cmd.getFlags() & UpdateCommand.IGNORE_AUTOCOMMIT) == 0) {
-        long currentTlogSize = getCurrentTLogSize();
         if (commitWithinSoftCommit) {
-          commitTracker.addedDocument(-1, currentTlogSize);
+          commitTracker.addedDocument(-1, this::getCurrentTLogSize);
           softCommitTracker.addedDocument(cmd.commitWithin);
         } else {
           softCommitTracker.addedDocument(-1);
-          commitTracker.addedDocument(cmd.commitWithin, currentTlogSize);
+          commitTracker.addedDocument(cmd.commitWithin, this::getCurrentTLogSize);
         }
       }
 
@@ -471,8 +490,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
         commitTracker.scheduleCommitWithin(commitTracker.getTimeUpperBound());
       }
 
-      long currentTlogSize = getCurrentTLogSize();
-      commitTracker.scheduleMaxSizeTriggeredCommitIfNeeded(currentTlogSize);
+      commitTracker.scheduleMaxSizeTriggeredCommitIfNeeded(this::getCurrentTLogSize);
 
       if (softCommitTracker.getTimeUpperBound() > 0) {
         softCommitTracker.scheduleCommitWithin(softCommitTracker.getTimeUpperBound());
@@ -555,12 +573,23 @@ public class DirectUpdateHandler2 extends UpdateHandler
     deleteByQueryCommandsCumulative.mark();
     boolean madeIt = false;
     try {
+      Query q = getQuery(cmd);
+
+      // Parts of the DBQ codepath run the query using a standard Lucene IndexSearcher, so block any
+      // queries that we know require a SolrIndexSearcher
+      final var unsupportedQDetector = new SolrSearcherRequirementDetector();
+      q.visit(unsupportedQDetector);
+      if (unsupportedQDetector.getRequiresSolrSearcher()) {
+        throw new SolrException(
+            SolrException.ErrorCode.BAD_REQUEST,
+            "Query [" + cmd.getQuery() + "] is not supported in delete-by-query operations");
+      }
+
       if ((cmd.getFlags() & UpdateCommand.IGNORE_INDEXWRITER) != 0) {
         if (ulog != null) ulog.deleteByQuery(cmd);
         madeIt = true;
         return;
       }
-      Query q = getQuery(cmd);
 
       boolean delAll = MatchAllDocsQuery.class == q.getClass();
 
@@ -633,7 +662,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
       }
       RefCounted<IndexWriter> iw = solrCoreState.getIndexWriter(core);
       try {
-        iw.get().addIndexes(mergeReaders.toArray(new CodecReader[mergeReaders.size()]));
+        iw.get().addIndexes(mergeReaders.toArray(new CodecReader[0]));
       } finally {
         iw.decref();
       }
@@ -721,6 +750,9 @@ public class DirectUpdateHandler2 extends UpdateHandler
 
       RefCounted<IndexWriter> iw = solrCoreState.getIndexWriter(core);
       try {
+        if (cmd.isClosingOnCommit()) {
+          core.readOnly = true;
+        }
         IndexWriter writer = iw.get();
         if (cmd.optimize) {
           writer.forceMerge(cmd.maxOptimizeSegments);
@@ -769,7 +801,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
           if (ulog != null) ulog.preSoftCommit(cmd);
           if (cmd.openSearcher) {
             core.getSearcher(true, false, waitSearcher);
-          } else {
+          } else if (!cmd.isClosingOnCommit()) {
             // force open a new realtime searcher so realtime-get and versioning code can see the
             // latest
             RefCounted<SolrIndexSearcher> searchHolder = core.openNewSearcher(true, true);
@@ -812,7 +844,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
       try {
         waitSearcher[0].get();
       } catch (InterruptedException | ExecutionException e) {
-        SolrException.log(log, e);
+        log.error("Exception waiting for searcher", e);
       }
     }
   }
@@ -954,8 +986,10 @@ public class DirectUpdateHandler2 extends UpdateHandler
 
             // todo: refactor this shared code (or figure out why a real CommitUpdateCommand can't
             // be used)
-            SolrIndexWriter.setCommitData(writer, cmd.getVersion(), null);
-            writer.commit();
+            if (shouldCommit(cmd, writer)) {
+              SolrIndexWriter.setCommitData(writer, cmd.getVersion(), cmd.commitData);
+              writer.commit();
+            }
 
             synchronized (solrCoreState.getUpdateLock()) {
               ulog.postCommit(cmd);
@@ -1038,7 +1072,7 @@ public class DirectUpdateHandler2 extends UpdateHandler
       // skips uniqueKey and _root_
       List<IndexableField> fields = cmd.makeLuceneDocForInPlaceUpdate().getFields();
       log.debug("updateDocValues({})", cmd);
-      writer.updateDocValues(updateTerm, fields.toArray(new Field[fields.size()]));
+      writer.updateDocValues(updateTerm, fields.toArray(new Field[0]));
 
     } else { // more normal path
